requirements.txt
huggingface-hub
numpy
pandas
pillow >= 9.5.0
simple-parsing >= 0.1.5
jax[cpu]
flax
einops
flax
jax
huggingface_hub
pandas
pillow
simple-parsing
